OCR/pyproject.toml
[tool.poetry]
package-mode = false
name = "ocr"
version = "0.1.0"
description = "Digitize scanned PDFs"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.4"
opencv-python = "^4.9.0.80"
levenshtein = "^0.25.1"
fastapi = {extras = ["standard"], version = "^0.112.1"}
pillow = "^10.3.0"
datasets = "^3.0.1"
tesserocr = "^2.7.1"

[tool.poetry.group.torch.dependencies]
transformers = {extras = ["torch"], version = "^4.45.1"}

[tool.poetry.group.dev.dependencies]
lxml = "^5.3.0"
docopt = "^0.6.2"
ruff = "^0.3.7"
pytest = "^8.1.1"
pytest-benchmark = "^4.0.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
main = "ocr.main:main"
api = "ocr.api:start"
build = "ocr.pyinstaller:install"

[tool.ruff]
line-length = 118
target-version = "py310"

[tool.ruff.lint]
select = ["D"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
# Ignore test directories
"tests/**" = ["D"]

# Ignore module-level init.py
"__init__.py" = ["D"]

# Ignore CLI entry points
"main.py" = ["D"]
"**_main.py" = ["D"]

# Ignore phdc_converter (already documented in a different style)
"ocr/services/phdc_converter/**" = ["D"]
